+++ packages/web/src/pages/upload-page/fields/availability/HiddenAvailabilityFields.module.css
